Code quality updates
