+++ src/PSRule.Common.props
@@ -3,7 +3,7 @@
   <!-- Project defaults -->
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
-    <LangVersion>9.0</LangVersion>
+    <LangVersion>12.0</LangVersion>
     <!-- <Nullable>enable</Nullable> -->
     <!-- <ImplicitUsings>enable</ImplicitUsings> -->
     <NeutralLanguage>en-US</NeutralLanguage>
