+++ src/PSRule.Rules.Azure/Data/AvailabilityZoneMapping.cs
@@ -3,24 +3,22 @@
 
 using Newtonsoft.Json;
 
-namespace PSRule.Rules.Azure.Data
+namespace PSRule.Rules.Azure.Data;
+
+/// <summary>
+/// An Azure Availability Zone mapping between location and zones.
+/// </summary>
+public sealed class AvailabilityZoneMapping
 {
+    /// <summary>
+    /// The location/ region where AZ is available.
+    /// </summary>
+    [JsonProperty(PropertyName = "l")]
+    public string Location { get; set; }
 
     /// <summary>
-    /// An Azure Availability Zone mapping between location and zones.
+    /// The zone names available at the location.
     /// </summary>
-    public sealed class AvailabilityZoneMapping
-    {
-        /// <summary>
-        /// The location/ region where AZ is available.
-        /// </summary>
-        [JsonProperty(PropertyName = "l")]
-        public string Location { get; set; }
-
-        /// <summary>
-        /// The zone names available at the location.
-        /// </summary>
-        [JsonProperty(PropertyName = "z")]
-        public string[] Zones { get; set; }
-    }
+    [JsonProperty(PropertyName = "z")]
+    public string[] Zones { get; set; }
 }
