+++ src/PSRule.Rules.Azure/Data/Template/BaseResourceValue.cs
@@ -1,24 +1,16 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace PSRule.Rules.Azure.Data.Template
+namespace PSRule.Rules.Azure.Data.Template;
+
+internal abstract class BaseResourceValue(string id, string name, string symbolicName)
 {
-    internal abstract class BaseResourceValue
-    {
-        protected BaseResourceValue(string id, string name, string symbolicName)
-        {
-            Id = id;
-            Name = name;
-            SymbolicName = symbolicName;
-        }
+    /// <inheritdoc/>
+    public string Id { get; } = id;
 
-        /// <inheritdoc/>
-        public string Id { get; }
+    /// <inheritdoc/>
+    public string Name { get; } = name;
 
-        /// <inheritdoc/>
-        public string Name { get; }
-
-        /// <inheritdoc/>
-        public string SymbolicName { get; }
-    }
+    /// <inheritdoc/>
+    public string SymbolicName { get; } = symbolicName;
 }
