+++ src/PSRule.Rules.Azure/Data/Template/DebugSymbol.cs
@@ -1,21 +1,14 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace PSRule.Rules.Azure.Data.Template
+namespace PSRule.Rules.Azure.Data.Template;
+
+/// <summary>
+/// A symbol used for debugging and runtime information.
+/// </summary>
+internal sealed class DebugSymbol(string name, string symbolName)
 {
-    /// <summary>
-    /// A symbol used for debugging and runtime information.
-    /// </summary>
-    internal sealed class DebugSymbol
-    {
-        public DebugSymbol(string name, string symbolName)
-        {
-            Name = name;
-            SymbolName = symbolName;
-        }
+    public string Name { get; } = name;
 
-        public string Name { get; }
-
-        public string SymbolName { get; }
-    }
+    public string SymbolName { get; } = symbolName;
 }
