+++ src/PSRule.Rules.Azure/Data/Template/DeploymentScope.cs
@@ -1,19 +1,18 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace PSRule.Rules.Azure.Data.Template
+namespace PSRule.Rules.Azure.Data.Template;
+
+/// <summary>
+/// The ARM scope of the deployment.
+/// </summary>
+internal enum DeploymentScope
 {
-    /// <summary>
-    /// The ARM scope of the deployment.
-    /// </summary>
-    internal enum DeploymentScope
-    {
-        ResourceGroup = 0,
+    ResourceGroup = 0,
 
-        Subscription = 1,
+    Subscription = 1,
 
-        ManagementGroup = 2,
+    ManagementGroup = 2,
 
-        Tenant = 3,
-    }
+    Tenant = 3,
 }
