+++ src/PSRule.Rules.Azure/Data/Template/DeploymentSymbol.cs
@@ -1,25 +1,19 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace PSRule.Rules.Azure.Data.Template
+namespace PSRule.Rules.Azure.Data.Template;
+
+internal abstract class DeploymentSymbol(string name)
 {
-    internal abstract class DeploymentSymbol
+    public string Name { get; } = name;
+
+    internal static IDeploymentSymbol NewArray(string name)
     {
-        protected DeploymentSymbol(string name)
-        {
-            Name = name;
-        }
+        return name == null ? null : new ArrayDeploymentSymbol(name);
+    }
 
-        public string Name { get; }
-
-        internal static IDeploymentSymbol NewArray(string name)
-        {
-            return name == null ? null : new ArrayDeploymentSymbol(name);
-        }
-
-        internal static IDeploymentSymbol NewObject(string name, ResourceValue resource = null)
-        {
-            return name == null ? null : new ObjectDeploymentSymbol(name, resource);
-        }
+    internal static IDeploymentSymbol NewObject(string name, ResourceValue resource = null)
+    {
+        return name == null ? null : new ObjectDeploymentSymbol(name, resource);
     }
 }
