+++ src/PSRule.Rules.Azure/Data/Template/IFunctionDescriptor.cs
@@ -1,21 +1,20 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace PSRule.Rules.Azure.Data.Template
+namespace PSRule.Rules.Azure.Data.Template;
+
+/// <summary>
+/// A common interface for defining and invoking ARM functions.
+/// </summary>
+internal interface IFunctionDescriptor
 {
     /// <summary>
-    /// A common interface for defining and invoking ARM functions.
+    /// The name of the function.
     /// </summary>
-    internal interface IFunctionDescriptor
-    {
-        /// <summary>
-        /// The name of the function.
-        /// </summary>
-        string Name { get; }
+    string Name { get; }
 
-        /// <summary>
-        /// Invoke the function.
-        /// </summary>
-        object Invoke(ITemplateContext context, DebugSymbol debugSymbol, ExpressionFnOuter[] args);
-    }
+    /// <summary>
+    /// Invoke the function.
+    /// </summary>
+    object Invoke(ITemplateContext context, DebugSymbol debugSymbol, ExpressionFnOuter[] args);
 }
