+++ src/PSRule.Rules.Azure/Data/Template/IResourceIdentity.cs
@@ -1,23 +1,22 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace PSRule.Rules.Azure.Data.Template
+namespace PSRule.Rules.Azure.Data.Template;
+
+internal interface IResourceIdentity
 {
-    internal interface IResourceIdentity
-    {
-        /// <summary>
-        /// The identifier of the resource.
-        /// </summary>
-        public string Id { get; }
+    /// <summary>
+    /// The identifier of the resource.
+    /// </summary>
+    public string Id { get; }
 
-        /// <summary>
-        /// The name of the resource.
-        /// </summary>
-        public string Name { get; }
+    /// <summary>
+    /// The name of the resource.
+    /// </summary>
+    public string Name { get; }
 
-        /// <summary>
-        /// The symbolic name of the resource.
-        /// </summary>
-        public string SymbolicName { get; }
-    }
+    /// <summary>
+    /// The symbolic name of the resource.
+    /// </summary>
+    public string SymbolicName { get; }
 }
