+++ src/PSRule.Rules.Azure/Data/Template/IUnknownMock.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+#nullable enable
+
+namespace PSRule.Rules.Azure.Data.Template;
+
+internal interface IUnknownMock : IMock
+{
+
+}
+
+#nullable restore
