+++ src/PSRule.Rules.Azure/Data/Template/TokenStream.cs
@@ -5,155 +5,154 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Text;
 
-namespace PSRule.Rules.Azure.Data.Template
+namespace PSRule.Rules.Azure.Data.Template;
+
+/// <summary>
+/// A stream of template expression tokens.
+/// </summary>
+[DebuggerDisplay("Current = (Type = {Current.Type}, Content = {Current.Content})")]
+internal sealed class TokenStream
 {
-    /// <summary>
-    /// A stream of template expression tokens.
-    /// </summary>
-    [DebuggerDisplay("Current = (Type = {Current.Type}, Content = {Current.Content})")]
-    internal sealed class TokenStream
+    private const char Comma = ',';
+    private const char Dot = '.';
+    private const char BackSlash = '\'';
+    private const char FunctionOpen = '(';
+    private const char FunctionClose = ')';
+    private const char IndexOpen = '[';
+    private const char IndexClose = ']';
+
+    private readonly List<ExpressionToken> _Token;
+
+    private int _Position;
+
+    internal TokenStream()
     {
-        private const char Comma = ',';
-        private const char Dot = '.';
-        private const char BackSlash = '\'';
-        private const char FunctionOpen = '(';
-        private const char FunctionClose = ')';
-        private const char IndexOpen = '[';
-        private const char IndexClose = ']';
+        _Token = [];
+    }
 
-        private readonly List<ExpressionToken> _Token;
+    internal TokenStream(IEnumerable<ExpressionToken> tokens)
+        : this()
+    {
+        foreach (var token in tokens)
+            Add(token);
+    }
 
-        private int _Position;
+    #region Properties
 
-        internal TokenStream()
+    public ExpressionToken Current => (_Token.Count <= _Position) ? null : _Token[_Position];
+
+    public int Count => _Token.Count;
+
+    #endregion Properties
+
+    public bool TryTokenType(ExpressionTokenType tokenType, out ExpressionToken token)
+    {
+        token = null;
+        if (Current == null || Current.Type != tokenType)
+            return false;
+
+        token = Pop();
+        return true;
+    }
+
+    public bool Skip(ExpressionTokenType tokenType)
+    {
+        if (Current == null || Current.Type != tokenType)
+            return false;
+
+        Pop();
+        return true;
+    }
+
+    public void SkipGroup()
+    {
+        if (!TryTokenType(ExpressionTokenType.GroupStart, out _))
+            return;
+
+        var inner = 0;
+        while (inner >= 0 && Count > 0)
         {
-            _Token = new List<ExpressionToken>();
-        }
+            if (Current.Type == ExpressionTokenType.GroupStart)
+                inner++;
 
-        internal TokenStream(IEnumerable<ExpressionToken> tokens)
-            : this()
-        {
-            foreach (var token in tokens)
-                Add(token);
-        }
-
-        #region Properties
-
-        public ExpressionToken Current => (_Token.Count <= _Position) ? null : _Token[_Position];
-
-        public int Count => _Token.Count;
-
-        #endregion Properties
-
-        public bool TryTokenType(ExpressionTokenType tokenType, out ExpressionToken token)
-        {
-            token = null;
-            if (Current == null || Current.Type != tokenType)
-                return false;
-
-            token = Pop();
-            return true;
-        }
-
-        public bool Skip(ExpressionTokenType tokenType)
-        {
-            if (Current == null || Current.Type != tokenType)
-                return false;
+            if (Current.Type == ExpressionTokenType.GroupEnd)
+                inner--;
 
             Pop();
-            return true;
         }
+    }
 
-        public void SkipGroup()
+    public void Add(ExpressionToken token)
+    {
+        _Token.Add(token);
+        _Position = _Token.Count - 1;
+    }
+
+    public ExpressionToken Pop()
+    {
+        if (Count == 0)
+            return null;
+
+        var token = _Token[_Position];
+        _Token.RemoveAt(_Position);
+        return token;
+    }
+
+    public void MoveTo(int position)
+    {
+        _Position = position;
+    }
+
+    internal ExpressionToken[] ToArray()
+    {
+        return _Token.ToArray();
+    }
+
+    internal string AsString()
+    {
+        var builder = new StringBuilder();
+        builder.Append(IndexOpen);
+        var group = 0;
+        ExpressionToken last = null;
+        for (var i = 0; i < _Token.Count; i++)
         {
-            if (!TryTokenType(ExpressionTokenType.GroupStart, out _))
-                return;
+            var current = _Token[i];
+            if (last != null && !(current.Type == ExpressionTokenType.Property || IsStartOrEndToken(current)) && !(last.Type == ExpressionTokenType.GroupStart || last.Type == ExpressionTokenType.IndexStart))
+                builder.Append(Comma);
 
-            var inner = 0;
-            while (inner >= 0 && Count > 0)
+            if (current.Type == ExpressionTokenType.Property)
+                builder.Append(Dot);
+            else if (current.Type == ExpressionTokenType.GroupStart)
             {
-                if (Current.Type == ExpressionTokenType.GroupStart)
-                    inner++;
-
-                if (Current.Type == ExpressionTokenType.GroupEnd)
-                    inner--;
-
-                Pop();
+                builder.Append(FunctionOpen);
+                group++;
             }
-        }
-
-        public void Add(ExpressionToken token)
-        {
-            _Token.Add(token);
-            _Position = _Token.Count - 1;
-        }
-
-        public ExpressionToken Pop()
-        {
-            if (Count == 0)
-                return null;
-
-            var token = _Token[_Position];
-            _Token.RemoveAt(_Position);
-            return token;
-        }
-
-        public void MoveTo(int position)
-        {
-            _Position = position;
-        }
-
-        internal ExpressionToken[] ToArray()
-        {
-            return _Token.ToArray();
-        }
-
-        internal string AsString()
-        {
-            var builder = new StringBuilder();
-            builder.Append(IndexOpen);
-            var group = 0;
-            ExpressionToken last = null;
-            for (var i = 0; i < _Token.Count; i++)
+            else if (current.Type == ExpressionTokenType.GroupEnd)
             {
-                var current = _Token[i];
-                if (last != null && !(current.Type == ExpressionTokenType.Property || IsStartOrEndToken(current)) && !(last.Type == ExpressionTokenType.GroupStart || last.Type == ExpressionTokenType.IndexStart))
-                    builder.Append(Comma);
-
-                if (current.Type == ExpressionTokenType.Property)
-                    builder.Append(Dot);
-                else if (current.Type == ExpressionTokenType.GroupStart)
-                {
-                    builder.Append(FunctionOpen);
-                    group++;
-                }
-                else if (current.Type == ExpressionTokenType.GroupEnd)
-                {
-                    builder.Append(FunctionClose);
-                    group--;
-                }
-                else if (current.Type == ExpressionTokenType.IndexStart)
-                    builder.Append(IndexOpen);
-                else if (current.Type == ExpressionTokenType.IndexEnd)
-                    builder.Append(IndexClose);
-                else if (current.Type == ExpressionTokenType.String)
-                    builder.Append(BackSlash);
-                else if (current.Type == ExpressionTokenType.Numeric)
-                    builder.Append(current.Value);
-
-                builder.Append(current.Content);
-                if (current.Type == ExpressionTokenType.String)
-                    builder.Append(BackSlash);
-
-                last = current;
+                builder.Append(FunctionClose);
+                group--;
             }
-            builder.Append(IndexClose);
-            return builder.ToString();
-        }
+            else if (current.Type == ExpressionTokenType.IndexStart)
+                builder.Append(IndexOpen);
+            else if (current.Type == ExpressionTokenType.IndexEnd)
+                builder.Append(IndexClose);
+            else if (current.Type == ExpressionTokenType.String)
+                builder.Append(BackSlash);
+            else if (current.Type == ExpressionTokenType.Numeric)
+                builder.Append(current.Value);
 
-        private static bool IsStartOrEndToken(ExpressionToken token)
-        {
-            return token.Type is ExpressionTokenType.GroupStart or ExpressionTokenType.GroupEnd or ExpressionTokenType.IndexStart or ExpressionTokenType.IndexEnd;
+            builder.Append(current.Content);
+            if (current.Type == ExpressionTokenType.String)
+                builder.Append(BackSlash);
+
+            last = current;
         }
+        builder.Append(IndexClose);
+        return builder.ToString();
+    }
+
+    private static bool IsStartOrEndToken(ExpressionToken token)
+    {
+        return token.Type is ExpressionTokenType.GroupStart or ExpressionTokenType.GroupEnd or ExpressionTokenType.IndexStart or ExpressionTokenType.IndexEnd;
     }
 }
