+++ src/PSRule.Rules.Azure/Data/Template/TypeDefinition.cs
@@ -3,30 +3,13 @@
 
 using Newtonsoft.Json.Linq;
 
-namespace PSRule.Rules.Azure.Data.Template
+namespace PSRule.Rules.Azure.Data.Template;
+
+internal sealed class TypeDefinition(TypePrimitive type, JObject definition, bool nullable) : ITypeDefinition
 {
-    internal interface ITypeDefinition
-    {
-        TypePrimitive Type { get; }
+    public TypePrimitive Type { get; } = type;
 
-        JObject Definition { get; }
+    public JObject Definition { get; } = definition;
 
-        bool Nullable { get; }
-    }
-
-    internal sealed class TypeDefinition : ITypeDefinition
-    {
-        public TypeDefinition(TypePrimitive type, JObject definition, bool nullable)
-        {
-            Type = type;
-            Definition = definition;
-            Nullable = nullable;
-        }
-
-        public TypePrimitive Type { get; }
-
-        public JObject Definition { get; }
-
-        public bool Nullable { get; }
-    }
+    public bool Nullable { get; } = nullable;
 }
