+++ src/PSRule.Rules.Azure/Data/Template/ValidationIssue.cs
@@ -1,19 +1,18 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace PSRule.Rules.Azure.Data.Template
-{
-    internal sealed class ValidationIssue
-    {
-        private readonly ValidationKind _Kind;
-        private readonly string _Name;
-        private readonly string _Message;
+namespace PSRule.Rules.Azure.Data.Template;
 
-        public ValidationIssue(ValidationKind kind, string name, string message)
-        {
-            _Kind = kind;
-            _Name = name;
-            _Message = message;
-        }
+internal sealed class ValidationIssue
+{
+    private readonly ValidationKind _Kind;
+    private readonly string _Name;
+    private readonly string _Message;
+
+    public ValidationIssue(ValidationKind kind, string name, string message)
+    {
+        _Kind = kind;
+        _Name = name;
+        _Message = message;
     }
 }
