+++ src/PSRule.Rules.Azure/Data/Template/ValidationKind.cs
@@ -1,12 +1,11 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace PSRule.Rules.Azure.Data.Template
-{
-    internal enum ValidationKind
-    {
-        None = 0,
+namespace PSRule.Rules.Azure.Data.Template;
 
-        Parameter = 1
-    }
+internal enum ValidationKind
+{
+    None = 0,
+
+    Parameter = 1
 }
