+++ src/PSRule.Rules.Azure/Runtime/IService.cs
@@ -3,13 +3,12 @@
 
 using System;
 
-namespace PSRule.Rules.Azure.Runtime
-{
-    /// <summary>
-    /// A context when running within PSRule.
-    /// </summary>
-    public interface IService : IDisposable
-    {
+namespace PSRule.Rules.Azure.Runtime;
 
-    }
+/// <summary>
+/// A context when running within PSRule.
+/// </summary>
+public interface IService : IDisposable
+{
+
 }
